trieda pre spravu okien
